+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- HTTP server pool examples: express-cluster, express-hybrid.
+
 ## [2.6.27] - 2023-08-15
 
 ### Fixed